refactor: split transactions.py into modular package
Break 3,591-line monolith into 11 focused modules:
- crud.py: list, create, view, edit, delete (497 lines)
- participants.py: add/remove participants (112 lines)
- api.py: JSON endpoints - search, signers, status, rentcast (274 lines)
- intake.py: questionnaire and document package generation (467 lines)
- documents.py: form filling (538 lines)
- signing.py: DocuSeal e-signature operations (840 lines)
- download.py: PDF download and print (413 lines)
- docuseal_admin.py: admin endpoints and webhook (241 lines)
- history.py: audit history API (110 lines)
- decorators.py: shared transactions_required decorator (20 lines)
- helpers.py: build_prefill_data, download_and_store functions (186 lines)

All 43 routes preserved with same URLs and url_for references.
No breaking changes to templates or other code.
